modified float to double

+++ simulation.cpp
@@ -51,7 +51,7 @@ valarray<double> f(double t, valarray<double> x) {
   for(int i=0; i<4; i++) {
     result[i] += q[i] * s_inv;
     for(int j=0; j<4; j++) {
-      float coef = pump[i][j] ? x[i]-x[j] >= 0 : -pump[j][i];
+      double coef = pump[i][j] ? x[i]-x[j] >= 0 : -pump[j][i];
       result[i] -= coef * s_inv  * sqrt(abs(x[i]-x[j]));
     }
     result[i] -= leakage[i] * s_inv * sqrt(abs(x[i]));
